Fix clippy warnings in TUI
- Add .await to tokio::sync::mpsc::Sender::send() calls in background.rs
- Use .first() instead of .get(0) in commands.rs
- Use strip_prefix instead of manual slicing in commands.rs
- Remove redundant if/else in input.rs
- Remove unnecessary cast in scrollbar.rs
- Simplify queue check in app.rs

+++ crates/sibyl-tui/src/widgets/scrollbar.rs
@@ -29,7 +29,7 @@ pub fn render_scrollbar(
     };
 
     for i in 0..scrollbar_height {
-        let y = area.y + 1 + i as u16;
+        let y = area.y + 1 + i;
         let x = area.x + area.width.saturating_sub(1);
 
         let char = if i as usize >= thumb_position && (i as usize) < thumb_position + thumb_height {
